fix: preserve full chain IDs in browser bindings

+++ helios-ts/src/ethereum.rs
@@ -174,8 +174,8 @@ impl EthereumClient {
     }
 
     #[wasm_bindgen]
-    pub fn chain_id(&self) -> u32 {
-        self.chain_id as u32
+    pub fn chain_id(&self) -> u64 {
+        self.chain_id
     }
 
     #[wasm_bindgen]
